+++ CHANGELOG.md
@@ -7,8 +7,9 @@ All notable changes to this project will be documented in this file. This projec
 ### Added
 - [#315](https://github.com/cloudcreativity/laravel-json-api/issues/315)
 Allow developers to use the exact JSON API field name as the relationship method name on their
-adapters. Although we recommend following the PSR1 standard of using camel case for method names,
-this does allow a developer to use snake case field names with snake case method names.
+adapters, plus for default conversion of names in include paths. Although we recommend following
+the PSR1 standard of using camel case for method names, this does allow a developer to use snake
+case field names with snake case method names.
 
 ## [1.0.1] - 2019-03-12